Fix: ContentLength type: `long`

+++ src/Postmark/Model/PostmarkInboundMessage.cs
@@ -152,7 +152,8 @@ namespace PostmarkDotNet
         public string Name { get; set; }
         public string Content { get; set; }
         public string ContentType { get; set; }
-        public int ContentLength { get; set; }
+        public long ContentLength { get; set; }
+
         public string ContentID { get; set; }
     }
